:art: pull out user generator into test-utils
